27. Mocking in Flask with Pytest/source_code/requirements.txt
connexion
pytest-flask
pytest-mock
requests
